src/poll_feed/requirements.txt
functions-framework==3.*
requests==2.26.0
google-cloud-secret-manager==2.16.4
google-cloud-pubsub==2.18.4
google-cloud-firestore==2.12.0